Added some code to help us resolve concrete types.
we can now resove any type which is registered to another type regardless of the generic parameters that are used.

This was achieved in the resolve method of the container, where we check if the source has generic types and if the map has those types.

then we find the generic type, and convert that into a closed type.

passing this closed type to the create instance method will allow us to return an instace of the class.

This is a very very very complex idea. Inversion of control. :/ Phew....